Use the configured asset scale, rather than ONE.
We can't fall back to lower scales, but if we configure a scale, then it's
saner to assume that all assets are going to be available at that scale.

Note: we don't use assetScale() here because that falls back to the scale
declared by Graphics, which might be automatically set to 2 if the game is
running on a HiDPI browser. But unless the game knows that it has 2x assets,
it's not safe to automatically use them in HiDPI mode, so tragically a game
will have to manually call setAssetScale() to make things work in HiDPI. Blah.