Change Assets.getBytes to use ByteBuffer.
This allows more flexibility under the hood to be efficient. On iOS we return a
ByteBuffer over the mmapped file, on Java we use FileChannel and a direct
ByteBuffer (which avoids extra copying).

This also includes an implementation of getBytes for the HTML5 backend (where
turning a TypedArray into a ByteBuffer is also natural and efficient). This
implementation only works on browsers that support TypedArray, but that's
better than not supporting it at all.

Thanks to @mbarbeaux for the motivation.