Add a #version header to our GLSL shaders.
This is slightly complicated by the fact that we want #version 100 for GLES2
but the desktop backend doesn't actually run on GLES, it runs on regular
OpenGL, and on the Mac at least, #version 100 is not supported so we have to
conditionally say we want #version 120 on that backend. Woo!