Totally revamped Java backend based on LWJGL3.
The SWT backend is currently not working due to an inability to set GL version
when creating a GL context, but work is underway by LWJGL3 people to replicate
the SWT/GL bridge and bring it kicking and screaming out of the last millenium.

Thanks to @tomfisher for getting the ball rolling on this.