Copy with disappearance of pyx4me.com.
We now avoid creating a symlink on Android, as we can simply refer
directly to the assets submodule subdir that contains our assets. On
robovm we use the antrun plugin to create our symlink, which only works
on non-Windows operating systems, but that's OK because you can only
build the robovm backend on a Mac anyway.