Make isLoaded mean "is successfully loaded."
This is going to be what we want in the vast majority of cases, and avoids
potential inconsistency in setSource() between sync and async images (thanks
Guillermo for pointing that out).

Our use of isLoaded() in ImageLayer is really just a (potentially pointless)
optimization, to avoid creating a lambda and passing it to the future, and
anyone else using isLoaded() is likely to be doing the same optimization, or at
least similarly checking both that the image is loaded and can be used (i.e. is
not in a loading failed state).