Expose createPath() and createGradient() on Graphics.
We'll keep them on Canvas as well because sometimes you have a Canvas reference
and don't have a Graphics reference. But sometimes you have the opposite, so no
need to force the creation of a dummy canvas just to get a Path or Gradient
object that you're going to use later when you *do* have a canvas.